pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "bahn_alarm"
version = "0.1.0"
description = ""
authors = [""]
maintainers = [""]


[tool.poetry.dependencies]
python = "^3.7"

fastapi = "0.88.0"
gunicorn = "^20.0.4"
uvicorn = "^0.20.0"
schiene = "^0.26"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
tortoise-orm = {extras = ["accel", "asyncpg"], version = "^0.19.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
aerich = "^0.7.1"
dramatiq = {extras = ["redis", "watch"], version = "^1.13.0"}
structlog = "^22.3.0"
better-exceptions = "^0.3.3"
rich = "^12.6.0"


[tool.poetry.dev-dependencies]
coverage = { version = "7.0.0", extras = ["toml"]}
black = "^23.1a1"
isort = { version = "^5.11.3", extras = ["pyproject"]}
requests = "^2.23.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"

[tool.black]
# https://github.com/psf/black
line-length = 88
target_version = ['py38']
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | venv
    | .venv
  )/
)
'''

[tool.isort]
# https://github.com/timothycrosley/isort
# https://github.com/timothycrosley/isort/wiki/isort-Settings
line_length = 88
indent = '    '
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0


[tool.coverage]
# https://github.com/nedbat/coveragepy
    [tool.coverage.run]

    source = ["app"]
    branch = true
    omit = ['']

    [tool.coverage.report]
    exclude_lines = [
        "# noqa",
        "raise NotImplementedError"
    ]

[tool.aerich]
tortoise_orm = "app.database.ORM_SETTINGS"
location = "./migrations"
src_folder = "./."
